Move FObasePage in FO directory
