+++ tests/UI/pages/FO/classic/cms/index.ts
@@ -1,5 +1,5 @@
 // Import pages
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 /**
  * CMS page, contains functions that can be used on the page
