+++ tests/UI/pages/FO/classic/delivery/index.ts
@@ -1,5 +1,5 @@
 // Import pages
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 /**
  * Delivery page, contains functions that can be used on the page
