+++ tests/UI/pages/FO/classic/emailSubscription/index.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 import {Page} from 'playwright';
 
 /**
