+++ tests/UI/pages/FO/classic/myAccount/merchandiseReturns.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 import {MerchandiseReturns as OrderMerchandiseReturns} from '@data/types/order';
