+++ tests/UI/pages/FO/classic/myAccount/orderDetails.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 import {MerchandiseProductReturn} from '@data/types/order';
