+++ tests/UI/pages/FO/classic/orderTracking/guestOrderTracking.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 /**
  * Guest order tracking page, contains functions that can be used on the page
