+++ tests/UI/pages/FO/classic/passwordReminder/index.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 import type {Page} from 'playwright';
