+++ tests/UI/pages/FO/classic/pricesDrop/index.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 /**
  * Prices drop page, contains functions that can be used on the page
