+++ tests/UI/pages/FO/classic/termsAndConditionsOfUse/index.ts
@@ -1,4 +1,4 @@
-import FOBasePage from '@pages/FO/classic/FObasePage';
+import FOBasePage from '@pages/FO/FObasePage';
 
 /**
  * Terms and conditions of use page, contains functions that can be used on the page
